@@ -8,7 +8,7 @@ readme = "README.md"
 packages = [{include = "grteamxaridbot"}]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.10"
 pyTelegramBotAPI = "^4.10.0"
 requests = "^2.28.2"
 openpyxl = "^3.1.1"
