@@ -1,4 +1,4 @@
 import logging
 
-logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.ERROR, filename='./../log.txt')
+logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.ERROR, filename='./log.txt')
 log = logging.getLogger(__name__)
